+++ Canvas - Copy/src/Canvas/Shapes/Oval.java
@@ -25,9 +25,9 @@ public class Oval extends Obj{
     @Override
     public void show(Graphics2D g2dBuffer){
         if (fill){
-            g2dBuffer.fillOval(-width/2, -length/2, width, length);
+            g2dBuffer.fillOval((int)-width/2, (int)-length/2, (int)width, (int)length);
         }else{
-            g2dBuffer.drawOval(-width/2, -length/2, width, length);
+            g2dBuffer.drawOval((int)-width/2, (int)-length/2, (int)width, (int)length);
         }
     }
 }
