+++ Canvas - Copy/src/Canvas/Shapes/Square.java
@@ -2,8 +2,8 @@ package Canvas.Shapes;
 import java.awt.*;
 
 public class Square extends Polygoni{
-    public Square(int X,int Y,int SideLength, Color Color,boolean tf){
-        super("Square",X,Y,new int []{0,SideLength,SideLength,0},new int []{0,0,SideLength,SideLength},Color,tf);
+    public Square(double X,double Y,double SideLength, Color Color,boolean tf){
+        super("Square",X,Y,new double []{0,SideLength,SideLength,0},new double []{0,0,SideLength,SideLength},Color,tf);
     }
     /**
      * Set the size of the square
